Fixes anonymous users not being able to view Custom Art images.
If an anonymous user uploaded a Custom Art image, it would be displayed
as a broken image in the browser.  uc_attribute_files module uses the
user_files module which places some security around uploaded files by
associating a file with a user.  When uploading an image, it was not
being associated with the anonymous user.

- When a file is uploaded, it is now associated with the anon user.
- Added a better message if we couldn't determine the uploaded filename.
- Uploading a file will also work now after adding to the cart failed
  it's validation.